# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon2026 Backend API
  version: 1.3.0
  description: |
    OpenAPI contract for the Hackathon2026 backend.

    Authentication model:
    - `POST /auth/psychologists/login` sets `session_id` and `csrf_token` cookies.
    - `POST /auth/admin/login` sets `admin_session_id` and `csrf_token` cookies.
    - Protected `GET` endpoints require the corresponding session cookie.
    - Protected non-`GET` endpoints also require header `X-CSRF-Token` with the same value as the `csrf_token` cookie.

    Practical notes for frontend integration:
    - Base URL in local development is `http://localhost:8080`.
    - Error responses use the unified shape `{ "message": "...", "field_errors": { ... } }`.
    - Browser CORS preflight is supported for configured `ALLOWED_ORIGINS`, including credentials and the `X-CSRF-Token` header.
    - Psychologist user payloads return `portal_access_until` and `blocked_until` as `null` when they are not set.
    - Psychologist user payloads may include computed `account_status` and `subscription_status`.
    - `GET /psychologists/tests` returns activity metrics such as started/completed counts and last activity timestamps for the “Мои опросники” dashboard.
    - Test creation defaults `status` to `published` when omitted.
    - Test update defaults `status` to `draft` when omitted.
    - Public test start requires `respondent_name` and `respondent_phone`.
...
    - In-progress public sessions expire after 1 hour of inactivity.
...
servers:
...
paths:
...
  /testdb:
    get:
      tags:
        - System
      summary: Database connectivity check
      operationId: testDatabase
      responses:
        '200':
          description: Database connection is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseStatusResponse'
        '503':
          description: Database connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/psychologists/login:
    post:
      tags:
        - Psychologist Auth
      summary: Login psychologist
      description: Creates a psychologist session and issues both session and CSRF cookies.
      operationId: loginPsychologist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PsychologistLoginInput'
      responses:
        '200':
          description: Authenticated psychologist session created
          headers:
            Set-Cookie:
              description: "Sets `session_id` and `csrf_token` cookies."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistAuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/psychologists/logout:
    post:
      tags:
        - Psychologist Auth
      summary: Logout psychologist
      operationId: logoutPsychologist
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Session cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/admin/login:
    post:
      tags:
        - Admin Auth
      summary: Login admin
      description: Creates an admin session and issues both session and CSRF cookies.
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginInput'
      responses:
        '200':
          description: Authenticated admin session created
          headers:
            Set-Cookie:
              description: "Sets `admin_session_id` and `csrf_token` cookies."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/admin/logout:
    post:
      tags:
        - Admin Auth
      summary: Logout admin
      operationId: logoutAdmin
      security:
        - adminSession: []
          csrfToken: []
      responses:
        '200':
          description: Session cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/me:
    get:
      tags:
        - Psychologist Workspace
      summary: Get psychologist workspace
      operationId: getPsychologistWorkspace
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Psychologist workspace with profile, card, and tests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistWorkspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/me/profile:
    get:
      tags:
        - Psychologist Workspace
      summary: Get psychologist profile
      operationId: getPsychologistProfile
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Psychologist profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Psychologist Workspace
      summary: Update psychologist profile
      operationId: updatePsychologistProfile
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdatePsychologistProfileInput'
      responses:
        '200':
          description: Updated psychologist profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/me/card:
    get:
      tags:
        - Psychologist Workspace
      summary: Get psychologist card
      operationId: getPsychologistCard
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Psychologist card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistCard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Psychologist Workspace
      summary: Update psychologist card
      operationId: updatePsychologistCard
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdatePsychologistCardInput'
      responses:
        '200':
          description: Updated psychologist card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistCard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/report-templates:
    post:
      tags:
        - Report Templates
      summary: Create report template
      description: Creates a reusable report template owned by the authenticated psychologist.
      operationId: createReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportTemplateInput'
      responses:
        '201':
          description: Report template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Report Templates
      summary: List report templates
      operationId: listReportTemplates
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Report templates created by the authenticated psychologist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/report-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags:
        - Report Templates
      summary: Get report template
      operationId: getReportTemplate
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Report template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Report Templates
      summary: Update report template
      operationId: updateReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportTemplateInput'
      responses:
        '200':
          description: Updated report template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Report Templates
      summary: Delete report template
      operationId: deleteReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Report template deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests:
    post:
      tags:
        - Tests
      summary: Create test
      description: Creates a test owned by the authenticated psychologist.
      operationId: createPsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestInput'
      responses:
        '201':
          description: Test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Tests
      summary: List tests
      description: Returns psychologist tests together with dashboard metrics used by the “Мои опросники” screen.
      operationId: listPsychologistTests
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Tests created by the authenticated psychologist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Test'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    get:
      tags:
        - Tests
      summary: Get test
      operationId: getPsychologistTest
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Test details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Tests
      summary: Update test
      operationId: updatePsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTestInput'
      responses:
        '200':
          description: Updated test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Tests
      summary: Delete test
      operationId: deletePsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Test deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/questions:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    post:
      tags:
        - Questions
      summary: Create question
      operationId: createPsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionInput'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Questions
      summary: List questions
      operationId: listPsychologistQuestions
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Questions for the selected test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
      - $ref: '#/components/parameters/QuestionIdParam'
    get:
      tags:
        - Questions
      summary: Get question
      operationId: getPsychologistQuestion
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Question details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Questions
      summary: Update question
      operationId: updatePsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionInput'
      responses:
        '200':
          description: Updated question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Questions
      summary: Delete question
      operationId: deletePsychologistQuestion
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    post:
      tags:
        - Tests
      summary: Publish test
      operationId: publishPsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Test published and public URL returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/results:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    get:
      tags:
        - Tests
      summary: List test submissions
      operationId: listPsychologistTestSubmissions
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Public test submissions for the selected test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PsychologistTestSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/results/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
      - $ref: '#/components/parameters/SubmissionSessionIdParam'
    get:
      tags:
        - Tests
      summary: Get test submission
      operationId: getPsychologistTestSubmission
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Test submission with answers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PsychologistTestSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/results/{sessionId}/report:
    parameters:
      - $ref: '#/components/parameters/SubmissionSessionIdParam'
      - $ref: '#/components/parameters/ReportFormatParam'
      - $ref: '#/components/parameters/ReportAudienceParam'
    get:
      tags:
        - Tests
      summary: Generate report for psychologist or client
      description: |
        Generates a report on demand for a completed submission.
        The report is never stored as a file on the server and is rebuilt for each request.
        By default `audience=psychologist` returns a technical report for the specialist.
        Use `audience=client` to generate the detailed client-facing report that the psychologist can later send manually to the client.
        Use `format=html` or `format=docx`.
      operationId: getPsychologistReportBySessionId
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Generated psychologist report
          content:
            text/html:
              schema:
                type: string
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}/formulas:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    post:
      tags:
        - Formula Rules
      summary: Create formula rule
      operationId: createFormulaRule
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFormulaRuleInput'
      responses:
        '201':
          description: Formula rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormulaRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Formula Rules
      summary: List formula rules
      operationId: listFormulaRules
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Formula rules for the selected test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FormulaRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /public/tests/{slug}:
    parameters:
      - $ref: '#/components/parameters/SlugParam'
    get:
      tags:
        - Public Tests
      summary: Get public test
      operationId: getPublicTest
      responses:
        '200':
          description: Public test with questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    psychologistSession:
      type: apiKey
      in: cookie
      name: session_id
      description: Psychologist session cookie returned by the psychologist login endpoint.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session_id
      description: Admin session cookie returned by the admin login endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Must match the `csrf_token` cookie value on protected non-GET requests."
...